Add files via upload
I have made a Leetmatric-app. When a user enter their username then all the data of that person will be displayed that how many questions that person solved on Leetcode
This project will tell the amount of Easy, Medium and Hard questions solved by user. In this project I used Leetcode API that fatch user data.